requirements.txt.txt
streamlit
gspread
pandas
unidecode
google-auth
